requirements.txt
requests
faker
pycryptodome